Move isolate_timestamp_in_release() to artcommon
