+++ artcommon/tests/test_util.py
@@ -3,7 +3,7 @@ import unittest
 import yaml
 
 from artcommonlib import util, build_util, release_util
-from artcommonlib.model import Model, MissingModel
+from artcommonlib.model import Model
 from artcommonlib.release_util import SoftwareLifecyclePhase
 from artcommonlib.util import deep_merge, isolate_major_minor_in_group
 
@@ -226,3 +226,32 @@ class TestSoftwareLifecyclePhase(unittest.TestCase):
         self.assertEqual(SoftwareLifecyclePhase.PRE_RELEASE, 0)
         self.assertEqual(SoftwareLifecyclePhase.EOL.value, 100)
         self.assertNotEqual(SoftwareLifecyclePhase.EOL.value, 101)
+
+    def test_isolate_timestamp_in_release(self):
+        actual = release_util.isolate_timestamp_in_release("foo-4.7.0-202107021813.p0.g01c9f3f.el8")
+        expected = "202107021813"
+        self.assertEqual(actual, expected)
+
+        actual = release_util.isolate_timestamp_in_release("foo-container-v4.7.0-202107021907.p0.g8b4b094")
+        expected = "202107021907"
+        self.assertEqual(actual, expected)
+
+        actual = release_util.isolate_timestamp_in_release("foo-container-v4.7.0-202107021907.p0.g8b4b094")
+        expected = "202107021907"
+        self.assertEqual(actual, expected)
+
+        actual = release_util.isolate_timestamp_in_release("foo-container-v4.8.0-202106152230.p0.g25122f5.assembly.stream")
+        expected = "202106152230"
+        self.assertEqual(actual, expected)
+
+        actual = release_util.isolate_timestamp_in_release("foo-container-v4.7.0-1.p0.g8b4b094")
+        expected = None
+        self.assertEqual(actual, expected)
+
+        actual = release_util.isolate_timestamp_in_release("foo-container-v4.7.0-202199999999.p0.g8b4b094")
+        expected = None
+        self.assertEqual(actual, expected)
+
+        actual = release_util.isolate_timestamp_in_release("")
+        expected = None
+        self.assertEqual(actual, expected)
